crm_system/pyproject.toml
[project]
name = "crm-system"
version = "0.1.0"
description = ""
authors = [
    {name = "evgenator",email = "zenyavolobuev@mail.ru"}
]
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-phonenumber-field[phonenumberslite] (>=8.2.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
